// swiyu-trust-management-ui/src/app/api/generated/model/non-compliant-actor.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {NonCompliantReasonText} from './non-compliant-reason-text';

export interface NonCompliantActor {
  id: string;
  did?: string;
  businessPartnerId?: string;
  flaggedAsNonCompliantAt: string;
  reason: NonCompliantReasonText;
}